# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Earn Daily API - Energy, Pending Coins, Confirmed Coins, withdrawals, and provider callbacks
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface InitUserRequest {
  deviceId: string;
}

export interface UserData {
  deviceId: string;
  coinsBalance: number;
  pkrBalance: number;
  energyBalance: number;
  pendingCoinsBalance: number;
  confirmedCoinsBalance: number;
  totalEarnedCoins?: number;
  lastCheckInTimestamp?: string | null;
  dailySpinsUsed: number;
  dailyScratchUsed: number;
  lastSpinResetDate?: string | null;
  lastScratchResetDate?: string | null;
  createdAt?: string;
}

export interface RewardResult {
  success: boolean;
  message: string;
  energyAwarded: number;
  balanceAfterEnergy: number;
  spinsLeft?: number | null;
  scratchLeft?: number | null;
}


export interface Withdrawal {
  withdrawalId: string;
  deviceId: string;
  paymentMethod: string;
  accountNumber: string;
  accountTitle: string;
  amountPKR: number;
  status: string;
  timestamp: string;
}

export type WithdrawalRequestPaymentMethod = typeof WithdrawalRequestPaymentMethod[keyof typeof WithdrawalRequestPaymentMethod];


export const WithdrawalRequestPaymentMethod = {
  Easypaisa: 'Easypaisa',
  JazzCash: 'JazzCash',
} as const;

export interface WithdrawalRequest {
  paymentMethod: WithdrawalRequestPaymentMethod;
  accountNumber: string;
  accountTitle: string;
  amountPKR: number;
}

export interface ProviderWebhookPayload {
  userId?: string;
  deviceId?: string;
  transactionId?: string;
  payoutUSD?: number;
  coins?: number;
  status?: string;
  signature?: string;
  secret?: string;
  [key: string]: unknown;
}