# lib/api-zod/src/generated/types/initUserRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Earn Daily API - Energy, Pending Coins, Confirmed Coins, withdrawals, and provider callbacks
 * OpenAPI spec version: 0.1.0
 */

export interface InitUserRequest {
  deviceId: string;
}