# lib/api-zod/src/generated/types/providerWebhookPayload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Earn Daily API - Energy, Pending Coins, Confirmed Coins, withdrawals, and provider callbacks
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderWebhookPayload {
  userId?: string;
  deviceId?: string;
  transactionId?: string;
  payoutUSD?: number;
  coins?: number;
  status?: string;
  signature?: string;
  secret?: string;
  [key: string]: unknown;
}